Added more tests and fixes

+++ StringManager/StringManager.Core/Models/Tone.cs
@@ -6,8 +6,8 @@
 
         public string Name { get; set; }
 
-        public int Frequency { get; set; }
+        public double Frequency { get; set; }
 
-        public int WaveLenght { get; set; }
+        public double WaveLenght { get; set; }
     }
 }
